requirements.txt
colorama
google-cloud-translate
h5py
hanziconv
imgcat
IPython
lxml
matplotlib
nltk
numba
numpy
pandas
pillow
pycantonese
pypinyin
pysubs2
python-dateutil
setuptools
readline
